pyproject.toml
[project]
name = "opinion-mining"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask-limiter>=3.10.1",
    "flask>=3.1.0",
    "pandas>=2.2.3",
    "pip>=25.0",
    "praw>=7.8.1",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    # "transformers>=4.48.2", if grok is not used
    "flask-wtf>=1.2.2",
    "flask-talisman>=1.1.0",
    "tabulate>=0.9.0",
    "psycopg2-binary>=2.9.10",
    "googlesearch-python>=1.3.0",
    "nltk>=3.9.1",
    # Remove keybert if not directly used
    # "keybert>=0.9.0",
    # "rank-bm25>=0.2.2",
    "openai>=1.66.2",
    "loguru>=0.7.3",
    "psutil>=6.1.1",
    "matplotlib>=3.10.5",
    "uvicorn>=0.38.0",
    "ipykernel>=7.1.0",
    "fastapi>=0.121.3",
    "python-multipart>=0.0.20",
    "transformers>=4.57.3",
    "torch>=2.9.1",
    "pmaw>=3.0.0",
    "ddgs>=9.9.2",
    "sentence-transformers>=5.1.2",
    "scalene>=1.5.55",
    "memory-profiler>=0.61.0",
    "memray>=1.19.1",
    "pyspellchecker>=0.8.4",
    "bitsandbytes>=0.49.0",
]


[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]